# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/cse441_project/ui/bookticket/chooseseat/ChooseSeatActivity.java
#	app/src/main/java/com/example/cse441_project/utils/FirebaseUtils.java

+++ app/src/main/java/com/example/cse441_project/ui/ggpay/viewmodel/CheckoutViewModel.java
@@ -83,7 +83,7 @@ public class CheckoutViewModel extends AndroidViewModel {
      * @param priceCents the price to show on the payment sheet.
      * @return a Task with the payment information.
      */
-    public Task<PaymentData> getLoadPaymentDataTask(final long priceCents) {
+    public Task<PaymentData> getLoadPaymentDataTask(final double priceCents) {
         JSONObject paymentDataRequestJson = PaymentsUtil.getPaymentDataRequest(priceCents);
         if (paymentDataRequestJson == null) {
             return null;
